restructuring game/ai functionalities
