+++ package.json
@@ -6,11 +6,11 @@
     "jth"
   ],
   "scripts": {
-    "build": "tsc && tsc-alias",
     "setup": "npm install --silent && npm run build",
     "start": "npm run build && node --no-deprecation js/index.js",
     "test": "npm run build && node --no-deprecation js/testing.js",
-    "reset": "npm run build && node --no-deprecation js/reset.js"
+    "reset": "npm run build && node --no-deprecation js/reset.js",
+    "build": "tsc && tsc-alias"
   },
   "dependencies": {
     "@inquirer/prompts": "^7.3.1",
