+++ source/types/ISettings.d.ts
