+++ source/utilities/CacheService.ts
@@ -1,5 +1,5 @@
 import { ITheme } from "@utilities/ITheme.js";
-import { IThemeOverride } from "@utilities/IThemeOverides.js";
+import { IThemeOverride } from "@utilities/IThemeOveride.js";
 import Config from "@utilities/Config.js";
 import {
   Dungeon,
@@ -7,21 +7,36 @@ import {
 } from "@utilities/DungeonService.js";
 import { getTerm, Language } from "@utilities/LanguageService.js";
 import { log } from "@utilities/LogService.js";
-import {
-  getSettingsData,
-  saveSettingsData,
-} from "@utilities/SettingsService.js";
+import { getDataFromFile, saveDataToFile } from "./StorageService.js";
+import { IGameState } from "@utilities/IGameState.js";
 
-const standardTheme = Config.STANDARD_THEME;
-const standardLanguage = Config.STANDARD_LANGUAGE;
-const standardPassword = Config.STANDARD_PASSWORD;
+// #region Initiate
+let settings = null;
+try {
+  settings = getDataFromFile("settings");
+} catch (error) {
+  log("Cache Service: " + error);
+}
+let gameState = null;
+try {
+  gameState = getDataFromFile("gameState");
+} catch (error) {
+  log("Cache Service: " + error);
+}
 
-let cachedDungeon: Dungeon = initiateDungeonMapWithHallways();
-let cachedLanguage: Language;
-let cachedTheme: ITheme;
-let cachedPassword: string;
-
-load();
+let cachedLanguage: Language = settings?.language || Config.STANDARD_LANGUAGE;
+let cachedTheme: ITheme = settings?.theme || Config.STANDARD_THEME;
+let cachedDungeon = initiateDungeonMapWithHallways();
+let cachedPassword: string = settings?.password || Config.STANDARD_PASSWORD;
+let cachedGameState: IGameState = gameState || {
+  theme: null,
+  narrativeHistory: [],
+  conversationHistory: [],
+  choices: [],
+  plotStage: 0,
+  plotSummary: "",
+};
+// #endregion
 
 // #region Getters
 export function getDungeon() {
@@ -38,6 +53,14 @@ export function getLanguage() {
 
 export function getTheme() {
   return cachedTheme;
+}
+
+export function getGameState() {
+  return cachedGameState;
+}
+
+export function getGameStateProperty<K extends keyof IGameState>(property: K) {
+  return cachedGameState[property];
 }
 // #endregion
 
@@ -62,6 +85,7 @@ export function setLanguage(language: Language): void {
 
 export function setTheme(theme: IThemeOverride) {
   log(`Cache Service: Theme set to ${theme.name.en}`);
+  const standardTheme = Config.STANDARD_THEME;
 
   cachedTheme = {
     name: { de: theme.name.de, en: theme.name.en },
@@ -82,24 +106,39 @@ export function setPassword(password: string) {
   save();
 }
 
-// #endregion
-function load() {
-  let settings = null;
-  try {
-    settings = getSettingsData();
-  } catch (error) {
-    log("Cache Service: " + error);
-  }
-  cachedLanguage = settings?.language || standardLanguage;
-  cachedTheme = settings?.theme || standardTheme;
-  cachedPassword = settings?.password || standardPassword;
+export function setGameState(gameState: IGameState) {
+  cachedGameState = gameState;
+  log("Cache Service: Game State updated");
+  save();
 }
 
+/**
+ * Sets a specific property in the game state.
+ * @param property A property of the game state. Like: "theme" or "narrativeHistory"
+ * @param value The value to set.
+ */
+export function setGameStateProperty<K extends keyof IGameState>(
+  property: K,
+  value: IGameState[K]
+) {
+  cachedGameState[property] = value;
+  log("Cache Service: Game State Property updated");
+  save();
+}
+// #endregion
+
+// TODO: Implement all necessary functions from gamestate.ts
+
+/**
+ * Saves the current settings and game state to the file system.
+ */
 function save() {
-  saveSettingsData({
+  saveDataToFile("settings", {
     language: cachedLanguage,
     theme: cachedTheme,
     password: cachedPassword,
   });
-  log("Cache Service: Settings synced");
+  saveDataToFile("gameState", cachedGameState);
+  log("Cache Service: Settings and GameState saved");
 }
+
