+++ source/utilities/Config.ts
@@ -2,8 +2,9 @@ import chalk from "chalk";
 import path from "path";
 import crypto from "crypto";
 import { ITheme } from "@utilities/ITheme.js";
-import ICharacterData from "@utilities/ICharacterData.js";
+import ICharacter from "@utilities/ICharacter.js";
 import { Language } from "./LanguageService.js";
+import { IAbility } from "@utilities/IAbility.js";
 
 // #region Paths
 // Base directory (assumes process.cwd() is the project root)
@@ -23,6 +24,12 @@ const SETTINGS_FILE = path.join(STORAGE_DIR, "settings.json");
 
 // Character data file path
 const CHARACTER_FILE = path.join(STORAGE_DIR, "character.json");
+
+// Character data file path
+const CONTEXT_FILE = path.join(STORAGE_DIR, "context.json");
+
+// Game state file path
+const GAME_STATE_FILE = path.join(STORAGE_DIR, "gamestate.json");
 
 // Animation frames file path (e.g., for attack animations)
 const ATTACK_FRAMES_FILE = path.join(STORAGE_DIR, "attackframes.json");
@@ -44,7 +51,22 @@ const STANDARD_THEME: ITheme = {
   backgroundColor: "#222222",
   errorColor: "#FF5555",
 };
-const STANDARD_CHARACTER: ICharacterData = {
+const STANDARD_LANGUAGE: Language = "de";
+const STANDARD_PASSWORD: string = crypto
+  .createHash("sha256")
+  .update("123")
+  .digest("hex");
+// #endregion
+
+// #region Character
+/**
+ * List of available character classes
+ */
+const CHARACTER_CLASSES = ["swordsman", "mage", "archer", "thief"];
+/**
+ * The default character data
+ */
+const START_CHARACTER: ICharacter = {
   name: "Hans",
   class: "swordsman",
   level: 4,
@@ -63,7 +85,10 @@ const STANDARD_CHARACTER: ICharacterData = {
   inventory: [],
   lastPlayed: new Date().toLocaleDateString("de-DE"),
 };
-const STANDARD_CHARACTER_STATS: Record<string, ICharacterData["abilities"]> = {
+/**
+ * The default character stats by class
+ */
+const START_CHARACTER_STATS: Record<string, ICharacter["abilities"]> = {
   swordsman: {
     maxhp: 20,
     strength: 5,
@@ -97,30 +122,94 @@ const STANDARD_CHARACTER_STATS: Record<string, ICharacterData["abilities"]> = {
     luck: 8,
   },
 };
-const STANDARD_LANGUAGE: Language = "de";
-const STANDARD_PASSWORD: string = crypto
-  .createHash("sha256")
-  .update("123")
-  .digest("hex");
+/**
+ * The default character abilities by class
+ */
+const START_CHARACTER_ABILITIES: Record<string, IAbility[]> = {
+  swordsman: [
+    {
+      name: "Power Strike",
+      manaCost: 0,
+      type: "attack",
+      multiplier: 1.2,
+      description: "A powerful melee attack that deals extra damage.",
+    },
+    {
+      name: "Battle Cry",
+      manaCost: 0,
+      type: "buff",
+      buffAmount: 2,
+      description: "A cry that temporarily boosts your strength.",
+    },
+  ],
+  mage: [
+    {
+      name: "Fireball",
+      manaCost: 3,
+      type: "attack",
+      multiplier: 1.5,
+      description: "Hurls a fiery ball that explodes on impact.",
+    },
+    {
+      name: "Healing Light",
+      manaCost: 2,
+      type: "heal",
+      healAmount: 8,
+      description: "Summons a gentle light to restore your health.",
+    },
+  ],
+  thief: [
+    {
+      name: "Backstab",
+      manaCost: 0,
+      type: "attack",
+      multiplier: 1.8,
+      description:
+        "A stealthy attack from behind that deals significant damage.",
+    },
+    {
+      name: "Smoke Bomb",
+      manaCost: 0,
+      type: "buff",
+      buffAmount: 3,
+      description: "Creates a smokescreen to help you evade attacks.",
+    },
+  ],
+};
 // #endregion
 
-// #region Choices
-const CHARACTER_CLASSES = ["swordsman", "mage", "archer", "thief"];
+// #region AI-Prompts
+const ORIGIN_STORIES = [""];
 // #endregion
 
 export default {
+  // Paths
   ROOT_DIR,
-  ENV_FILE,
   STORAGE_DIR,
+  ENV_FILE,
   LOG_FILE,
   SETTINGS_FILE,
   CHARACTER_FILE,
+  CONTEXT_FILE,
+  GAME_STATE_FILE,
   ATTACK_FRAMES_FILE,
+
+  // Strings
   SELECT_SEPARATOR,
+
+  // Standard Settings
   STANDARD_THEME,
-  STANDARD_CHARACTER,
-  CHARACTER_CLASSES,
-  STANDARD_CHARACTER_STATS,
   STANDARD_LANGUAGE,
   STANDARD_PASSWORD,
+
+  // Character
+  START_CHARACTER,
+  START_CHARACTER_STATS,
+  START_CHARACTER_ABILITIES,
+
+  // Choices
+  CHARACTER_CLASSES,
+
+  // AI-Prompts
+  ORIGIN_STORIES,
 };
