+++ source/utilities/GameService.ts
@@ -5,17 +5,16 @@ import {
   generateChatNarrative,
   generateEnemyFromNarrative,
   ChatCompletionRequestMessage,
-} from "./aiAssistant.js";
+} from "@utilities/AIService.js";
 import { rollDice } from "@utilities/DiceService.js";
-import { GameState } from "./gameState.js";
-import { promptForChoice } from "./gameMaster.js";
 import { log, LogTypes } from "@utilities/LogService.js";
-import { runCombat } from "./combat.js";
 import { generateRandomItem } from "@utilities/ItemGenerator.js";
 import { saveGameState, loadGameState } from "@utilities/SaveLoadService.js";
 import { getStartingItems } from "@utilities/InventoryService.js";
-import { saveCharacterData } from "@utilities/CharacterService.js";
 import { getLanguage } from "@utilities/CacheService.js";
+import { getDataFromFile, saveDataToFile } from "@utilities/StorageService.js";
+import { GameState } from "src/gameState.js";
+import { runCombat } from "src/combat.js";
 
 /**
  * Displays a persistent status bar showing key character stats.
@@ -116,7 +115,7 @@ Combat & Special Encounters:
 - Do not provide in-game choices during combat or dungeon encounters.
 
 Player Actions & In-Game Syntax:
-- The player’s in-game actions must be enclosed in curly braces {like this}. Only perform dice rolls or game mechanics when the player uses the correct syntax.
+- The player's in-game actions must be enclosed in curly braces {like this}. Only perform dice rolls or game mechanics when the player uses the correct syntax.
 - In-character dialogue must be enclosed in quotation marks "like this".
 - Out-of-character instructions will be provided in angle brackets <like this>.
 - For non-combat scenes, always end with exactly three numbered choices enclosed in curly braces (e.g.,
@@ -290,7 +289,7 @@ Please respond in clear, concise ${getLanguage}.
             );
             characterData.inventory.push(newItem);
           }
-          saveCharacterData(characterData);
+          saveDataToFile("character", characterData);
           await pauseForReflection("Press Enter to continue your journey...");
         }
       } else if (narrative.toLowerCase().includes("roll a d20")) {
@@ -334,8 +333,7 @@ Please respond in clear, concise ${getLanguage}.
  * then entering the main campaign loop.
  */
 export async function startCampaign(): Promise<void> {
-  const { getCharacterData } = await import("@utilities/CharacterService.js");
-  const characterData = getCharacterData();
+  const characterData = getDataFromFile("character");
   if (!characterData) {
     log(
       "No character data found. Please create a character first.",
@@ -352,4 +350,48 @@ export async function startCampaign(): Promise<void> {
   }
   const gameState = new GameState();
   await campaignLoop(gameState, characterData);
+}
+
+/**
+ * Extracts option lines from the narrative using a regex in multiline mode.
+ * If at least three numbered options are found, they are used.
+ * Otherwise, a default set of options is provided.
+ * "Return to main menu" is always included.
+ */
+export async function promptForChoice(narrative: string): Promise<string> {
+  // Regex to capture lines that start with a number and a period.
+  const optionRegex = /^\s*\d+\.\s+(.*)$/gm;
+  const options: string[] = [];
+  let match: RegExpExecArray | null;
+
+  while ((match = optionRegex.exec(narrative)) !== null) {
+    if (match[1] && match[1].trim().length > 0) {
+      options.push(match[1].trim());
+    }
+  }
+
+  // If less than three options were found, fallback to default choices.
+  if (options.length < 3) {
+    options.length = 0; // Clear any partial results.
+    options.push("🛡️  Option 1: Proceed further into the ruins");
+    options.push("🔍  Option 2: Examine your surroundings");
+    options.push("📦  Option 3: Check your inventory");
+  }
+
+  // Ensure "Return to main menu" is always added.
+  if (
+    !options.some((opt) => opt.toLowerCase().includes("return to main menu"))
+  ) {
+    options.push("🔙 Return to main menu");
+  }
+
+  const { selectedOption } = await inquirer.prompt([
+    {
+      type: "list",
+      name: "selectedOption",
+      message: "👉 Choose an option:",
+      choices: options,
+    },
+  ]);
+  return selectedOption;
 }
